yaml utils: move OctInt from meta
Also add a unit test for dumping yaml from OctInt
